backend/requirements.txt
# Backend requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
pillow==10.2.0
scikit-learn==1.4.0
requests==2.31.0
